FLoops.jl

FLoops.FLoops — Module

FLoops: fold for humans™

FLoops.jl provides a macro @floop. It can be used to generate a fast generic iteration over complex collections.

See FoldsCUDA.jl for GPU support.

Usage

Sequential (single-thread) loops

Simply wrap a for loop and its initialization part by @floop:

julia> using FLoops  # exports @floop macro

julia> @floop begin
           s = 0
           for x in 1:3
               s += x
           end
       end
       s
6

For more examples, see sequential loops tutorial.

Parallel loop

Parallel loops require additional syntax @reduce.

julia> @floop for (x, y) in zip(1:3, 1:2:6)
           a = x + y
           b = x - y
           @reduce(s += a, t += b)
       end
       (s, t)
(15, -3)

For more examples, see parallel loops tutorial.

How it works

@floop works by converting the native Julia for loop syntax to foldl defined by Transducers.jl. Unlike foldl defined in Base, foldl defined by Transducers.jl is powerful enough to cover the for loop semantics and more.